move creation of s3 connection to a factory function
